requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
# used older version to avoid rust error
pydantic==1.10.13
python-dotenv==1.0.0
itsdangerous==2.1.2
email-validator==2.1.0 